Moves the mesh io functions into the semantically similar modules
